Update angle threshold check
It looks like the updateNeeded() method assumes that the angle() function returns the angle between the rotation (which seems resonable :)), but after studying the source code of TF I see that it actually returns the half angle. See https://github.com/ros/geometry/blob/noetic-devel/tf/include/tf/LinearMath/Quaternion.h#L211.

Doubling this half angle will make the republisher actually publish according to the angle threshold that I would expect as an user.

This would be a breaking change from the current behavior that has been around for a long time. However the _worst_ consequence seems to be that the republisher will emit faster.